Simple Collision Detection
Impletented a simple way we can do collision detection and added a small example in the player object.

Basically you just implement the Box2dCollisionHandler Interface on any objects that you add to the world that you want to handle collision for.

Then it is a matter of just using the category data(or type) to filter out results and react accordingly.

Also fixed a bug where the small player could not go through small passage ways.