Added Exit Object
The exit now works. right now it is set to complete the level with 1 player but can be set to require both players to touch it at the same time.

the level complete send the to the current level again but plan to make it set it to GameMain global "m_stage" and "m_next_level", which is tracked in main layer.

it needs a graphic and a fade animation still.