Added trampolines!
Added a basic trampoline(only works with players right now). But i should be able to expand it to work with other objects at a later time like blocks or something else.

Also added some updates to the exit(finish level) mechanic. Now the Both players need to reach the end in order to complete the level.

WARNING: the exit is a sensor so the player can go through it at the moment and fall off the map. If this happens simply press T to reload the level.