SWITCHES!
added a very scalable switch implementation.

Switch objects must have unique names.
The name of the switch does not matter as long as it is unique.

this is becuase switch triggers are now handled through an addapter class. This allows us to set multiple switches for a single switch and even have switches that toggle switches. The possibilities are ENDLESS lol.

Custom properties:
 --Switch
----toggle(bool):determines if the switch is able to toggle on and off
----sensitive(bool):determines if the switch requires a body to stay on
----startOn(bool):determine if the switch starts on(currently unused/untested)
----visible(bool):determines if the switch is visible.

--Platform:
----switches(String):comma delimited string that contains all the names of the switches that activate/trigger this object. [NOTICE:Any object may have switches attached to it. You must implement the SwitchHandler Interface and instantiate the switches(String) array in order to filter switch sources]