One way platforms
Implemented one way platforms. if the object is under the platform then it will not trigger collision with the platform. This works for everything that can collide with the platform including the brain/blocks and players.

Everything is hardcoded into the collisionAdapter but minor changes should be easy enough to implement.

One idea for players to be able to go under a platform by pressing a key such as down while on the platform is to have the player m_body transformed a single tenth lower in order to make the platform pass collision.