Platform HotFix
Fixed a bug where the player would fall through a platform when going up.

Also added "oneWay" property to platform that now indicates if the platform is a one way platform. If the property is ommited then the default value is true.

Added a simple way to go to next level by pressing the Q key.