Levels 1-5
Changes --
-levels 1-5
-old levels set to stage 2
-exit now sizes correctly
-changed world timestep to accomodate slower devices

Made some pretty simple levels and fixed some of the magic numbers in the map loading algorithm so that they are a bit easier to read.

I tried to make the levels as interesting as possible with the difficulty and constraints in mind(easy with only spikes).

I tested them and they dont seem too hard to complete . And i had someone else with only knowledge of the controls test them as well with relative ease.