Added Smasher
Added smasher and a DEBUG variable in GameOff class. By setting this to true you load the debug test level of your choice and enable the box2d debugrenderer.

Smasher Properties:
-acceleration = how fast this smasher travels when SMASHING ! :D
-riseSpeed = how fast to ruse back up once done smashing
-switches = the switches that trigger this smasher(pref invisible sensitive ones).

The  Smasher ONLY GOES DOWN for now. The size is set to the size of the object in Tiled, so change to your liking :)